+++ hi-ml-azure/src/health_azure/himl.py
@@ -28,8 +28,8 @@ from azureml.train.hyperdrive import HyperDriveConfig
 
 from health_azure.utils import (create_python_environment, create_run_recovery_id, _find_file,
                                 is_run_and_child_runs_completed, is_running_in_azure_ml, register_environment,
-                                run_duration_string_to_seconds,
-                                to_azure_friendly_string, RUN_CONTEXT, get_workspace, PathOrString)
+                                run_duration_string_to_seconds, to_azure_friendly_string, RUN_CONTEXT, get_workspace,
+                                PathOrString)
 from health_azure.datasets import (DatasetConfig, StrOrDatasetConfig, _input_dataset_key, _output_dataset_key,
                                    _replace_string_datasets)